uchar

+++ src/frontend/imageprocessorwrapper.h
@@ -7,11 +7,11 @@
 
 class ImageProcessorWrapper {
   private:
-    static std::vector<std::vector<std::vector<uint>>> rgbImageToVector(const QImage &);
+    static std::vector<std::vector<std::vector<uchar>>> rgbImageToVector(const QImage &);
     static std::vector<std::vector<uint>> imageToVector(const QImage &);
     static void vectorToImage(QImage &dest, const std::vector<std::vector<uint>> &src);
     static void vectorToRgbImage(QImage &dest,
-                                 const std::vector<std::vector<std::vector<uint>>> &src);
+                                 const std::vector<std::vector<std::vector<uchar>>> &src);
 
   public:
     static void invertColor(QImage &);
